New hook to handle the current scale from store
